Reorganize code structure and create Dockerfile
The reorganization includes moving fake_the_spire files to src/ directory and updating relevant file paths accordingly. The addition includes the creation of a Dockerfile setup for building and running the application in a container. Additionally, certain files in the test/ directory were deleted and changes were made to the specifications in the cards and enemies data files.
