--- fake_the_spire/data/enemies.toml
@@ -1,6 +1,0 @@
-[enemies]
-[enemies.louse]
-hp = 6
-max_hp = 6
-actions = ["attack 5"]
-valid_floors = [1, 2, 3]
